+++ src/plugins/estree.js
@@ -1,4 +1,4 @@
-import * as estreePrinter from "../language-js/printer-estree.js";
+import * as estreePrinter from "../language-js/printer.js";
 import jsLanguages from "../language-js/languages.evaluate.js";
 import * as estreeJsonPrinter from "../language-json/printer-estree-json.js";
 import jsonLanguages from "../language-json/languages.evaluate.js";
